Update macos_install.sh
Created a check to see if homebrew is installed, if it is not exit, to be installed manually (Note: Requires Sudo and we should not run LoLLMs with Sudo permission for script).

Check if GCC is installed in homebrew and update or install depending on status.

Created INSTALL_DIR variable that way all Directory settings are now based on the Install directory instead of using relative paths.