Update README.md
Updated the required commands in the *manual installation instructions* section in order to allow the ollama_ai bindings to locate the lollms.databases module it is trying to use.  Without
```bash
pip install -e lollms_core
```
present, there will only be a loose error saying that it could not find the module with no real way to determine the cause.  These instructions appear in the ManualInstallation.md file, but that is not mentioned in this README.md, nor is it clear that one should run pip install -e lollms_core to resolve the problem.